+++ src/assets/css/Card/AsideCard.css
@@ -5,7 +5,7 @@
     cursor: pointer;
     padding: 6px 12px;
     margin: 5px;
-    border-radius: 4px;
+    border-radius: 3px;
     background-color: #091e420a;
     color: #FFF;
 }
