+++ src/assets/css/Card/Checklist.css
@@ -4,7 +4,7 @@
     align-items: center;
     justify-content:space-between;
     padding: 5px;
-    border-radius: 5px;
+    border-radius: 3px;
 
 
 }
