+++ src/assets/css/Card/Members.css
@@ -16,7 +16,7 @@
     background-color: #ffffff;
     z-index: 5;
     padding: 10px;
-    border-radius: 6px;
+    border-radius: 3px;
     border: solid 1px rgba(186, 183, 183, 0.545)
 }
 
@@ -57,7 +57,7 @@
     cursor: pointer;
     padding: 10px 12px;
     margin: 5px;
-    border-radius: 4px;
+    border-radius: 3px;
     background-color: #091e420a;
     transition-property: background-color;
     transition-duration: 0.4s;
