+++ src/assets/css/Card/ModalCard.css
@@ -15,7 +15,7 @@
     height: fit-content;
     position: relative;
     margin: 100px 200px;
-    border-radius: 10px;
+    border-radius: 3px;
     background-color: #ebecf0;
     display: flex;
     flex-direction: column;
