+++ src/assets/css/Card/SelectDate.css
@@ -16,7 +16,7 @@
     background-color: #ffffff;
     z-index: 5;
     padding: 10px;
-    border-radius: 6px;
+    border-radius: 3px;
     border: solid 1px rgba(186, 183, 183, 0.545)
 }
 
@@ -49,7 +49,7 @@
 
 .modal__dates__savebttn, .modal__dates__removebttn {
     color: #ffffff;
-    border-radius: 5px;
+    border-radius: 3px;
     cursor: pointer;
     padding: 10px 6px;
     margin-bottom: 8px;
@@ -118,7 +118,7 @@
     flex-wrap: nowrap;
     margin-top: 3px;
     padding: 7px;
-    border-radius: 4px;
+    border-radius: 3px;
     align-items: flex-end;
     justify-content: space-between;
     cursor: pointer;
