+++ src/components/ShowedCard/ModalCard.jsx
@@ -9,7 +9,6 @@ const ModalCard = (props) => {
     const context = useAppContext()
     const [comments, setComments] = useState([])
     const [checkList, setCheckList] = useState([])
-    const [members, setMembers]= useState([])
 
     const getComments = async () => {
         try{
@@ -34,22 +33,10 @@ const ModalCard = (props) => {
             console.log("error. failed to load data.");
         }
     }
-    const getMembers = async () => {
-        try{
-            const resp = await cardsAPI.getMembers(context.keys, props.currentCard.idBoard);
-            if (resp.status===200){    
-                setMembers(resp.data);
-            }
-        } catch (error) {
-            console.log(error);
-            console.log("error. failed to load data.");
-        }
-    }
 
     useEffect(() => {
         getComments();
         getChecklist();
-        getMembers()
     }
         , [])
 
@@ -98,8 +85,7 @@ const ModalCard = (props) => {
                                 setCheckList={setCheckList} 
                                 checkList={checkList} 
                                 setCurrentCard={props.setCurrentCard}
-                                members={members}
-                                setMembers={setMembers}/>
+                               />
                         </div>
                     </div>
